+++ contrib/PanopticSeg/paddlepanseg/models/__init__.py
@@ -14,3 +14,5 @@
 
 from .backbones import *
 from .losses import *
+from .panoptic_deeplab import PanopticDeepLab
+from .mask2former import Mask2Former
