+++ contrib/PanopticSeg/paddlepanseg/models/mask2former/common.py
