+++ contrib/PanopticSeg/paddlepanseg/transforms/generate_targets/maskformer.py
