--- contrib/PanopticSeg/projects/README.md
@@ -1,2 +1,0 @@
-TODO:
-- project template
