--- contrib/PanopticSeg/projects/mask2former/run.py
@@ -1,29 +1,0 @@
-#!/usr/bin/env python
-
-# Copyright (c) 2021 PaddlePaddle Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import os.path as osp
-import sys
-
-from paddlepanseg.run_project import main
-
-# Project directory
-PROJ_DIR = osp.abspath(osp.dirname(__file__))
-# Project name (name of project directory)
-PROJ_NAME = osp.basename(PROJ_DIR)
-
-# By default run the global `main()`
-# Note that customized training/evaluation scripts can be configured here
-main([PROJ_NAME] + sys.argv[1:])
