+++ contrib/PanopticSeg/test_tipc/configs/panoptic_deeplab/train_infer_python.txt
@@ -14,7 +14,7 @@ train_infer_img_dir:test_tipc/data/coco/val_list.txt
 null:null
 ##
 trainer:norm
-norm_train:projects/panoptic_deeplab/run.py train --do_eval --log_iters 5
+norm_train:tools/train.py --do_eval --log_iters 5
 pact_train:null
 fpgm_train:null
 distill_train:null
@@ -30,7 +30,7 @@ null:null
 --model_path:adaptive
 --input_shape:-1 3 1024 1024
 --config:adaptive
-norm_export:projects/panoptic_deeplab/run.py export
+norm_export:tools/export.py
 quant_export:null
 fpgm_export:null
 distill_export:null
@@ -48,7 +48,7 @@ infer_model:null
 --image_path:test_tipc/data/coco/val_list.txt
 --save_log_path:null
 --benchmark:True
-inference:projects/panoptic_deeplab/run.py infer
+inference:deploy/python/infer.py
 infer_export:null
 infer_quant:False
 --model_name:panoptic_deeplab
