+++ contrib/PanopticSeg/tools/analyze_model.py
@@ -15,7 +15,6 @@
 # limitations under the License.
 
 import argparse
-import os
 import io
 import sys
 import contextlib
@@ -27,14 +26,12 @@ from paddle.hapi.dynamic_flops import (count_parameters, register_hooks,
                                        count_io_info)
 from paddleseg.utils import get_sys_env, logger, op_flops_funs
 
-from paddlepanseg.project_manager import work_on_project
 from paddlepanseg.cvlibs import Config
 
 
 def parse_args():
     parser = argparse.ArgumentParser(description="Model analysis")
     parser.add_argument('--config', help="Config file.", type=str)
-    parser.add_argument('--proj', help="Project name.", type=str)
     parser.add_argument(
         '--input_shape',
         nargs='+',
@@ -203,25 +200,12 @@ def analyze(args):
     custom_ops = {paddle.nn.SyncBatchNorm: op_flops_funs.count_syncbn}
     inputs = paddle.randn(args.input_shape)
 
-    if args.proj is not None:
-        config = os.path.abspath(args.config)
-        # NOTE: From #2806 PaddleSeg eagerly builds datasets during config 
-        # construction. We set `switch_dir` to True such that datasets can be 
-        # correctly built with the relative paths specified in the config file.
-        with work_on_project(args.proj, switch_dir=True):
-            cfg = Config(config)
-            dynamic_flops(
-                cfg.model,
-                inputs,
-                custom_ops=custom_ops,
-                num_levels=args.num_levels)
-    else:
-        cfg = Config(args.config)
-        dynamic_flops(
-            cfg.model,
-            inputs,
-            custom_ops=custom_ops,
-            num_levels=args.num_levels)
+    cfg = Config(args.config)
+    dynamic_flops(
+        cfg.model,
+        inputs,
+        custom_ops=custom_ops,
+        num_levels=args.num_levels)
 
 
 if __name__ == '__main__':
